backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
pydantic
ecdsa
environs
asyncpg
python-multipart